refactor: extract ScratchpadItem to its own file (INV-18)
