+++ apps/frontend/src/components/layout/sidebar/stream-item.tsx
@@ -1,18 +1,12 @@
 import { useRef, type ReactNode, type RefObject } from "react"
-import { Archive, Bell, FileEdit, Hash, Lock, MessageSquareText, MoreHorizontal, Settings, User } from "lucide-react"
+import { Bell, FileEdit, Hash, Lock, MessageSquareText, MoreHorizontal, Settings, User } from "lucide-react"
 import { Link } from "react-router-dom"
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu"
+import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu"
 import { Button } from "@/components/ui/button"
 import { MentionIndicator } from "@/components/mention-indicator"
 import { RelativeTime } from "@/components/relative-time"
 import { getThreadRootContext } from "@/components/thread/breadcrumb-helpers"
-import { useActors, useStreamOrDraft } from "@/hooks"
+import { useActors } from "@/hooks"
 import { useSidebar } from "@/contexts"
 import { useStreamSettings } from "@/components/stream-settings/use-stream-settings"
 import { cn } from "@/lib/utils"
@@ -22,8 +16,9 @@ import { useUrgencyTracking } from "./use-urgency-tracking"
 import { truncateContent } from "./utils"
 import { StreamTypes, Visibilities, type AuthorType, type StreamWithPreview } from "@threa/types"
 import type { StreamItemData, UrgencyLevel } from "./types"
+import { ScratchpadItem } from "./scratchpad-item"
 
-function UrgencyStrip({ urgency }: { urgency: UrgencyLevel }) {
+export function UrgencyStrip({ urgency }: { urgency: UrgencyLevel }) {
   return (
     <div
       className="w-1 flex-shrink-0 rounded-l-lg transition-colors duration-300"
@@ -38,7 +33,7 @@ interface StreamItemAvatarProps {
   badge?: { icon: typeof Hash; color: string } | null
 }
 
-function StreamItemAvatar({ icon, className, badge }: StreamItemAvatarProps) {
+export function StreamItemAvatar({ icon, className, badge }: StreamItemAvatarProps) {
   return (
     <div
       className={cn(
@@ -68,7 +63,7 @@ interface StreamItemPreviewProps {
   showPreviewOnHover: boolean
 }
 
-function StreamItemPreview({ preview, getActorName, compact, showPreviewOnHover }: StreamItemPreviewProps) {
+export function StreamItemPreview({ preview, getActorName, compact, showPreviewOnHover }: StreamItemPreviewProps) {
   if (!preview?.content) return null
 
   return (
@@ -87,7 +82,7 @@ function StreamItemPreview({ preview, getActorName, compact, showPreviewOnHover 
   )
 }
 
-function StreamItemContextMenu({ children }: { children: ReactNode }) {
+export function StreamItemContextMenu({ children }: { children: ReactNode }) {
   const { setMenuOpen } = useSidebar()
 
   return (
@@ -257,108 +252,6 @@ export function StreamItem({
         >
           <Settings className="mr-2 h-4 w-4" />
           Settings
-        </DropdownMenuItem>
-      </StreamItemContextMenu>
-    </Link>
-  )
-}
-
-interface ScratchpadItemProps {
-  workspaceId: string
-  stream: StreamItemData
-  isActive: boolean
-  unreadCount: number
-  mentionCount: number
-  showUrgencyStrip?: boolean
-  /** Show compact view (title only, no preview) */
-  compact?: boolean
-  /** Show preview on hover when compact (only works with compact=true) */
-  showPreviewOnHover?: boolean
-  /** Reference to scroll container for position tracking */
-  scrollContainerRef?: RefObject<HTMLDivElement | null>
-}
-
-function ScratchpadItem({
-  workspaceId,
-  stream: streamWithPreview,
-  isActive,
-  unreadCount,
-  mentionCount,
-  showUrgencyStrip = true,
-  compact = false,
-  showPreviewOnHover = false,
-  scrollContainerRef,
-}: ScratchpadItemProps) {
-  const { stream, isDraft, archive } = useStreamOrDraft(workspaceId, streamWithPreview.id)
-  const { getActorName } = useActors(workspaceId)
-  const { openStreamSettings } = useStreamSettings()
-  const itemRef = useRef<HTMLAnchorElement>(null)
-  const hasUnread = unreadCount > 0
-
-  const currentDisplayName = stream?.displayName ?? streamWithPreview.displayName ?? null
-  const name = currentDisplayName || streamFallbackLabel("scratchpad", "sidebar")
-  const preview = streamWithPreview.lastMessagePreview
-
-  useUrgencyTracking(itemRef, streamWithPreview.id, streamWithPreview.urgency, scrollContainerRef)
-
-  const handleArchive = async () => {
-    await archive()
-  }
-
-  return (
-    <Link
-      ref={itemRef}
-      to={`/w/${workspaceId}/s/${streamWithPreview.id}`}
-      className={cn(
-        "group relative flex items-stretch rounded-lg text-sm transition-colors",
-        isActive ? "bg-primary/10" : "hover:bg-muted/50",
-        hasUnread && !isActive && "bg-primary/5 hover:bg-primary/10"
-      )}
-    >
-      {showUrgencyStrip && <UrgencyStrip urgency={streamWithPreview.urgency} />}
-
-      <div className="flex items-center gap-2.5 flex-1 min-w-0 px-2 py-2">
-        <StreamItemAvatar icon={<FileEdit className="h-3.5 w-3.5" />} className="bg-primary/10 text-primary" />
-
-        <div className="flex flex-col flex-1 min-w-0 gap-0.5">
-          <div className="flex items-center gap-2 pr-8">
-            <span className={cn("truncate text-sm", hasUnread ? "font-semibold" : "font-medium")}>
-              {name}
-              {isDraft && <span className="ml-1.5 text-xs text-muted-foreground font-normal">(draft)</span>}
-            </span>
-            <MentionIndicator count={mentionCount} className="ml-auto" />
-          </div>
-          <StreamItemPreview
-            preview={preview}
-            getActorName={getActorName}
-            compact={compact}
-            showPreviewOnHover={showPreviewOnHover}
-          />
-        </div>
-      </div>
-
-      <StreamItemContextMenu>
-        {!isDraft && (
-          <DropdownMenuItem
-            onClick={(e) => {
-              e.stopPropagation()
-              openStreamSettings(streamWithPreview.id)
-            }}
-          >
-            <Settings className="mr-2 h-4 w-4" />
-            Settings
-          </DropdownMenuItem>
-        )}
-        {!isDraft && <DropdownMenuSeparator />}
-        <DropdownMenuItem
-          onClick={(e) => {
-            e.stopPropagation()
-            void handleArchive()
-          }}
-          className="text-destructive"
-        >
-          <Archive className="mr-2 h-4 w-4" />
-          {isDraft ? "Delete" : "Archive"}
         </DropdownMenuItem>
       </StreamItemContextMenu>
     </Link>
